Melhorias no script de desinstalação e adição de argumentos de exec

+++ uninstall.sh
@@ -1,2 +1,12 @@
+echo "Checking root permissions..."
+
+if [ "$EUID" -ne 0 ]; then
+    echo "Please run as root."
+    exit 1
+fi
+
+echo "Removing files"
 rm -rf /opt/tfiles
+
+echo "Removing binaries"
 rm /usr/bin/tfiles
